requirements.txt
python-telegram-bot==21.10
openai==1.64.0
python-dotenv
grobid-client-python==0.0.9
black
flake8
isort
pre-commit
xmltodict==0.14.2
